[Clean fluid] Clean hash, grid_sampler, log_loss, bilinear_tensor_product. (#48411)
* Clean fliud hash

* clean fluid grid_sampler

* clean log_loss

* Move bilinear_tensor_product from fluid to static

* Fix unitests when remove log_loss

* Fix bug when move bilinear_tensor_product

* fix test_fleet_nocvm_1.py

* Add bilinear_tensor_product into all list

* Fix code style

* Fix comments in bilinear_tensor_product

* Fix comments in bilinear_tensor_product

* Fix comments
