+++ python/paddle/fluid/tests/unittests/test_fleet_rolemaker.py
@@ -16,6 +16,7 @@
 import os
 import unittest
 
+import paddle
 import paddle.fluid.incubate.fleet.base.role_maker as role_maker
 
 
@@ -97,7 +98,7 @@ class TestCloudRoleMaker(unittest.TestCase):
                 append_batch_size=False,
             )
             label_cast = fluid.layers.cast(label, dtype='float32')
-            cost = fluid.layers.log_loss(fc, label_cast)
+            cost = paddle.nn.functional.log_loss(fc, label_cast)
         try:
             adam = fluid.optimizer.Adam(learning_rate=0.000005)
             adam = fleet.distributed_optimizer(adam)
