+++ python/paddle/fluid/tests/unittests/test_fleet_unitaccessor.py
@@ -16,6 +16,8 @@
 import os
 import unittest
 
+import paddle
+
 
 class TestFleet1(unittest.TestCase):
     """
@@ -73,7 +75,7 @@ class TestFleet1(unittest.TestCase):
                 append_batch_size=False,
             )
             label_cast = fluid.layers.cast(label, dtype='float32')
-            cost = fluid.layers.log_loss(fc, label_cast)
+            cost = paddle.nn.functional.log_loss(fc, label_cast)
 
         strategy = {}
         strategy["embedding"] = {}
