+++ python/paddle/fluid/tests/unittests/test_hash_op.py
@@ -17,8 +17,6 @@ import unittest
 import numpy as np
 from op_test import OpTest
 
-import paddle.fluid as fluid
-
 
 class TestHashOp(OpTest):
     def setUp(self):
@@ -118,45 +116,6 @@ class TestHashOp3(TestHashOp):
 
     def test_check_output(self):
         self.check_output()
-
-
-class TestHashOpError(unittest.TestCase):
-    def test_errors(self):
-        with fluid.program_guard(fluid.Program(), fluid.Program()):
-            input_data = np.random.randint(0, 10, (8, 1)).astype("int32")
-
-            def test_Variable():
-                # the input type must be Variable
-                fluid.layers.hash(input=input_data, hash_size=2**32)
-
-            self.assertRaises(TypeError, test_Variable)
-
-            def test_type():
-                # dtype must be int32, int64.
-                x2 = fluid.layers.data(
-                    name='x2', shape=[1], dtype="float32", lod_level=1
-                )
-                fluid.layers.hash(input=x2, hash_size=2**32)
-
-            self.assertRaises(TypeError, test_type)
-
-            def test_hash_size_type():
-                # hash_size dtype must be int32, int64.
-                x3 = fluid.layers.data(
-                    name='x3', shape=[1], dtype="int32", lod_level=1
-                )
-                fluid.layers.hash(input=x3, hash_size=1024.5)
-
-            self.assertRaises(TypeError, test_hash_size_type)
-
-            def test_num_hash_type():
-                # num_hash dtype must be int32, int64.
-                x4 = fluid.layers.data(
-                    name='x4', shape=[1], dtype="int32", lod_level=1
-                )
-                fluid.layers.hash(input=x4, hash_size=2**32, num_hash=2.5)
-
-            self.assertRaises(TypeError, test_num_hash_type)
 
 
 if __name__ == "__main__":
